Move tests for ExecCommands to separate file
